Finished subset sum greedy

+++ src/SubsetSumApprox.java
@@ -17,10 +17,10 @@ public interface SubsetSumApprox {
      * Returns the residue, the absolute value of the difference between
      * the sum of the subset and the desired sum.
      *
-     * @param multiset List of integers in the multiset S
+     * @param multiset array of integers in the multiset S
      * @param sum int indicating the specified sum k
      *
      * @return double - the residue
      */
-    double subsetResidue(List<Integer> multiset, int sum);
+    int subsetResidue(List<Integer> multiset, int sum);
 }
